+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @uigen-dev/cli
 
+## 0.16.0
+
+### Minor Changes
+
+- Added the electron Target
+
+### Patch Changes
+
+- Updated dependencies
+  - @uigen-dev/config-gui@0.16.0
+  - @uigen-dev/react@0.16.0
+  - @uigen-dev/core@0.16.0
+
 ## 0.15.0
 
 ### Minor Changes